move seaice helper functions into common
